main/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.9
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.3.3
idna==2.6
mongoengine==0.9.0
pymongo==2.7
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
